requirements.txt
click==8.0.3
lxml==4.6.4
requests==2.26.0
python-crontab==2.6.0